pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wakeflow"
version = "1.0.2"
description = "Generate and manipulate semi-analytic models of planet wakes"
readme = "README.md"
authors = [{ name = "Thomas Hilder", email = "tom.hilder.dlhp@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization"
]
keywords = ["astronomy", "astrophysics", "research", "protoplanets", "analytic", "modelling", "planet wake", "protoplanetary", "discs", "disks"]
dependencies = [
    "numpy >= 1.21.1",
    "matplotlib >= 3.5.0",
    "astropy >= 5.0",
    "scipy >= 1.8.0",
    "setuptools >= 61.0.0",
    "pyyaml >= 6.0",
    "tqdm >= 4.64.0",
    "pytest >= 7.1.0"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/TomHilder/wakeflow"